server/python/requirements.txt
Flask
psycopg2-binary==2.9.9
python-dotenv
pytest
flask-jwt-extended
bcrypt